requirements.txt
python-telegram-bot==21.5
twilio==9.3.0
